requirements.txt
tk
tkinterdnd2
sv_ttk
pillow
cryptography
ttkthemes
